chore(events): final-review minor cleanups
- ClipCutter sweeps orphaned staging/ dirs from unclean shutdowns
- lifecycle.py serializes embeddings as explicit little-endian float32
  (matches Go's binary.LittleEndian decode)
- recorder.go comments the keyless-message drop as malformed handling
- Thumb handler returns 500 on DB error instead of folding into 404
- cosineSim test covers non-multiple-of-4 byte lengths

+++ face-service/face_service/lifecycle.py
@@ -110,7 +110,8 @@ class LifecycleEmitter:
             msg["started_ts"] = st.started_ts
             msg["ended_ts"] = ts
             if st.crop_embedding is not None:
-                emb = np.asarray(st.crop_embedding, dtype=np.float32)
+                # Explicit little-endian: the Go side decodes with binary.LittleEndian.
+                emb = np.asarray(st.crop_embedding, dtype="<f4")
                 msg["embedding_b64"] = base64.b64encode(emb.tobytes()).decode("ascii")
             else:
                 msg["embedding_b64"] = None
